Refactor after rebase
